+++ wayland-feather-shot(1)/wayland-feather-shot/src/wayland_feather_shot/scrollcap/__init__.py
@@ -1,1 +1,0 @@
-"""Scroll capture helpers."""